Fixed my Cats Calendar.
It was trying to access a calendar the current google account did not
have access to. Research told me it throws a 404 when you try to list
event of a calendar that never existed or the account cant access. Added
code to reset GPC if account cant see current calendar.